Backend/pyproject.toml
[tool.poetry]
name = "asknust"
version = "0.1.0"
description = ""
authors = ["Samama251251 <samama251251@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = {extras = ["standard"], version = "^0.115.7"}
uvicorn = "^0.32.0"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
pypdf = "^5.1.0"
langchain-pinecone = "^0.2.0"
langchain-openai = "^0.2.8"
pypdf2 = "^3.0.1"
langchain-google-genai = "^2.0.9"
langchain-mistralai = "^0.2.5"
motor = "^3.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
